修改logSaplingInterval 类型

+++ common/java/common-web/src/main/java/com/welab/wefe/common/web/api/base/Api.java
@@ -81,7 +81,7 @@ public @interface Api {
      * 大于0：
      * 在周期内仅输出一次完整的 api 响应结果，其它响应省略输出。
      */
-    int logSaplingInterval() default 0;
+    long logSaplingInterval() default 0L;
 
     /**
      * forward matching uri
